Revert "Organización de estructura y link .env"
