+++ .env
@@ -1,0 +1,9 @@
+# Configuración de Neon PostgreSQL
+DATABASE_URL=postgresql://neondb_owner:npg_fu4XCqL1NQxV@ep-sparkling-math-ad1478s1-pooler.c-2.us-east-1.aws.neon.tech/neondb?sslmode=require&channel_binding=require
+
+# Variables individuales (opcional, para mayor flexibilidad)
+DB_HOST=ep-sparkling-math-ad1478s1-pooler.c-2.us-east-1.aws.neon.tech
+DB_PORT=5432
+DB_NAME=neondb
+DB_USERNAME=neondb_owner
+DB_PASSWORD=npg_fu4XCqL1NQxV
