--- .gitignore
@@ -1,78 +1,0 @@
-# Byte-compiled / optimized / DLL files
-__pycache__/
-*.py[cod]
-*$py.class
-
-# C extensions
-*.so
-
-# Virtual environment
-venv/
-.env/
-.venv/
-env/
-ENV/
-*.env
-
-# Jupyter Notebook checkpoints
-.ipynb_checkpoints
-
-# VS Code config
-.vscode/
-
-# PyCharm
-.idea/
-
-# mypy
-.mypy_cache/
-
-# Pyre type checker
-.pyre/
-
-# Pytype
-.pytype/
-
-# Coverage reports
-htmlcov/
-.coverage
-.coverage.*
-.cache
-nosetests.xml
-coverage.xml
-*.cover
-*.py,cover
-
-# Pytest
-.pytest_cache/
-
-# SQLAlchemy / Alembic (migraciones)
-alembic/versions/*.pyc
-alembic/versions/__pycache__/
-alembic.ini
-
-# Logs
-*.log
-
-# Distribution / packaging
-.Python
-build/
-develop-eggs/
-dist/
-downloads/
-eggs/
-.eggs/
-lib/
-lib64/
-parts/
-sdist/
-var/
-*.egg-info/
-.installed.cfg
-*.egg
-
-# MacOS
-.DS_Store
-
-# Windows
-Thumbs.db
-ehthumbs.db
