+++ Entities/genero.py
@@ -7,7 +7,7 @@ from pydantic import BaseModel, Field, validator
 from sqlalchemy.sql import func
 from datetime import datetime
 
-from ..database.config import Base  # Ajusta a tu ruta real
+from ..Database.config import Base  # Ajusta a tu ruta real
 
 
 class Genero(Base):
