+++ Model/Citas.py
@@ -3,7 +3,7 @@ from sqlalchemy.orm import relationship
 from pydantic import BaseModel, Field
 from datetime import datetime
 from typing import Optional
-from ..database.config import Base
+from ..Database.config import Base
 from uuid import UUID
 
 class Citas(Base):
