--- ORM/requirements.txt
@@ -1,5 +1,0 @@
-sqlalchemy==2.0.23
-psycopg2-binary==2.9.9
-asyncpg==0.29.0
-alembic==1.13.1
-python-dotenv==1.0.0
